[fold] rename `MemoryFactory` -> `MemDBFactory`
