+++ src/test/shamap/common.h
@@ -57,7 +57,7 @@ public:
         , j_(j)
     {
         Section testSection;
-        testSection.set("type", "testmemory");
+        testSection.set("type", "memory");
         testSection.set("path", "SHAMap_test");
         db_ = NodeStore::Manager::instance().make_Database(
             megabytes(4), scheduler_, 1, testSection, j);
